InMemory client session
